size_t/ptrdiff_t transition part 4: libast regex and string sublibraries
This is the fourth of the thickfold patch series, which enables ksh93
to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The libast string library.
- The libast regex engine.
- The libcmd paste builtin.
- A few scattered bits of miscellaneous code using stresc (which now
  returns a ptrdiff_t result). Full changes for other libcmd builtins
  and pty are currently held for later parts of thickfold.

Change in the number of warnings on Linux when compiling with clang using
-Wsign-compare -Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion:
3,084 => 2,474 => 37 (progression from part 3 => part 4 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libast/comp/wordexp.c
@@ -45,7 +45,7 @@ static int	sh_unquote(char* string)
 		return sp-string;
 	if((dp=sp) > string && sp[-1]=='$')
 	{
-		int n=stresc(sp+1);
+		ptrdiff_t n=stresc(sp+1);
 		/* copy all but trailing ' */
 		while(--n>0)
 			*dp++ = *++sp;
